Move edit code to subfolder in Media & Text block
