+++ packages/block-library/src/media-text/edit/media-container-icon.js
